refactor: move createCheckMessageOptions to validation-helpers [deps-change]
- Extracted createCheckMessageOptions function to a new validation-helpers.ts file
- Updated references in index.ts to use the new helper function
- Improved code organization and separation of concerns
